fix warning

+++ stepmania/src/ScreenEvaluation.cpp
@@ -211,7 +211,7 @@ void ScreenEvaluation::Init()
 		if( PREFSMAN->m_iScoringType == PrefsManager::SCORING_5TH )
 		{
 			const int ScoreBonuses[] = { 10000000, 10000000, 1000000, 100000, 10000, 1000, 100 };
-			if( grade[p] < ARRAYSIZE(ScoreBonuses) )
+			if( grade[p] < (int) ARRAYSIZE(ScoreBonuses) )
 			{
 				g_CurStageStats.iBonus[p] += ScoreBonuses[(int)grade[p] ];
 				stageStats.iBonus[p] += ScoreBonuses[(int)grade[p] ];
